Create GermanTrafficSignRecognition.py
German Traffic Sign Recognition

Data set Link

https://www.kaggle.com/meowmeowmeowmeowmeow/gtsrb-german-traffic-sign

The German Traffic Sign Recognition Benchmark is a computer vision dataset that consists of over 50,000 images of traffic signs, with 43 different classes. The dataset is used to evaluate the performance of image classification algorithms on real-world traffic sign recognition tasks.

The objective of this project is to build a machine learning model that can accurately classify the different types of traffic signs in the dataset. The model should be able to identify the correct class of a given image, even if the image is blurry, occluded, or taken from a different angle or lighting condition.
The project will involve several steps, including data preprocessing, feature extraction, model selection and training, and model evaluation. The deliverables of the project will be a trained model that can classify traffic signs with high accuracy, and a report or presentation that documents the approach, the results, and the limitations of the model.